[API] Refactor promotion coupon state processors to correct namespaces and service IDs
