+++ src/Sylius/Bundle/ApiBundle/StateProcessor/Admin/Promotion/PromotionCoupon/PersistProcessor.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\StateProcessor\Admin\PromotionCoupon;
+namespace Sylius\Bundle\ApiBundle\StateProcessor\Admin\Promotion\PromotionCoupon;
 
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\Metadata\Post;
